fix monad behavior of Escape.
The monad behavior of Escape built into >>= wasn't all the behavior
of escape.  Consequently, if you initalized an "Escape" yourself
or tried to use escape as a control structure in a do block there were
more than a few gotchas.  This fixes the behavior of bind so that when
bind is applied to an Escape, it carries over the filter from the Ok.

this simplied the implementation of escape and allowed writing of an
escape' control structure that lets you exit your WebT response building
do block immediately.