more polishing of the WebT/ServerPartT change.
This change generalizes all the guards and request
manipulators (like ok, dir, and path) so they work
as you'd expect inside our monad stack.  There are a 
few annoying parts but the essentials are all there.

This also adds versions of the guards that can be used
inside a do block, and adds substantially to the documentation.