add support for Transfer-Encoding: chunked, output
By default we now use, Transfer-Encoding: chunked, for sending the
Response. This means that we do not have to force the lazy ByteString
to calculate the Content-Length header.

For files sent via SendFile, chunk mode is not used. You can also use
the normal Content-Length method on a per Response basis by setting
the RsFlag.