Has inotify observe the folder containing a file, rather than the file itself. Otherwise inotify would stop listening at the file when editing it with vim.
Vim deletes the file and moves a separate copy of the file onto the original location. This caused inotify to stop looking at the file.

This patch also has happstack-plugins follow a sourcefile with compile errors on the first try to compile it. Previously, if the file had compiling errors the first time it was loaded by the application, happstack-plugins would not reload it when the file is fixed. This second fix required switching from using IORef to using MVars in order to allow IO operations (registering the broken source file) while updating the mapping of the observed files.