happstack-server 6.2.1: change how we detect and report a failure to call decodeBody In response to this error report:
http://www.haskell.org/pipermail/web-devel/2011/002051.html

RqEnv now has the type:

type RqEnv = ([(String, Input)], Maybe [(String, Input)], [(String, Cookie)])

The body inputs are now a Maybe value. 
  Nothing = no attempt has been made to decode the body
  Just i  = the results of attempting to decode the body

in bodyInput (called by decodeBody), we set rqBodyInputs to [] if
bodyInput is called, but the request did not contain any
form-data. (Before we left the MVar empty).

We now only report a failure to call decodeBody if a look function is
called that attempts to find values in the body but decodeBody has
never been called. You can avoid this error by calling decodeBody or
by using the queryString filter.