Has happstack-plugins remain in a consistent state in the presence of asynchronous exceptions while rebuilding plugins.
The following code 

> do addSymbol ph fp sym
>   rebuild ph fp True

was problematic when an exception were thrown during the evaluation of the
second line. The internal state would reference a symbol which would be never
loaded.

For this reason, the code was modified to:

> bracketOnError
>   (addSymbol ph fp sym)
>   (const$ deleteSymbol ph fp sym)
>   (const$ rebuild ph fp True)

which after executed, it either leaves a loaded symbol in the internal state,
or does not leave the symbol if the attempt to load it was interrupted by an
exception.