Prevents plugins from being unloaded in the presence of compilation errors.
This translates with a change in the signature of funcTH' from
funcTH :: PluginHandle -> Name -> IO (Either Errors a)
to
funcTH' :: PluginHandle -> Name -> [String] -> IO (Errors,Maybe a)

where the value a is always returned even in the presence of errors.
The only case in which Nothing is returned is when the module has not 
been loaded yet.